requirements.txt
flask
flask-cors
pre-commit
requests
elasticsearch
python-dotenv
pydantic